done movieController
