+++ src/controllers/page.js
@@ -2,20 +2,17 @@ import {
   NUMBER_SHOW_FILMS,
   NUMBER_SHOW_TOP_RATED_FILMS,
   NUMBER_SHOW_MOST_COMMENTED_FILMS,
-  KEY_CODE_ESCAPE,
   render,
   unrender
-} from "./utils";
-import Films from "./films";
-import NoFilms from "./no-flims";
-import Card from "./film-card";
-import Popup from "./popup";
-import Comment from "./comment";
-import BtnMore from "./btn-more";
-import Menu from "./menu";
-import Sort from "./sort";
+} from "../components/utils";
+import Films from "../components/films";
+import NoFilms from "../components/no-flims";
+import BtnMore from "../components/btn-more";
+import Menu from "../components/menu";
+import Sort from "../components/sort";
+import MovieController from "./movie";
 
-export default class PageController {
+export default class Page {
   constructor(container, cards) {
     this._container = container;
     this._cards = cards;
@@ -93,54 +90,8 @@ export default class PageController {
   }
 
   _renderFilm(filmCard, container) {
-    const footerElement = document.querySelector(`footer`);
-    const card = new Card(filmCard);
-    const popup = new Popup(filmCard);
-
-    const onEscKeyDown = (evt) => {
-      if (evt.keyCode === KEY_CODE_ESCAPE) {
-        unrender(popup.getElement());
-        popup.removeElement();
-        document.removeEventListener(`keydown`, onEscKeyDown);
-      }
-    };
-
-    const onClickCard = () => {
-      render(footerElement, popup.getElement(), `afterend`);
-
-      const commentsContainer = popup.getElement().querySelector(`.film-details__comments-list`);
-      filmCard.comments.forEach((comment) => {
-        render(commentsContainer, new Comment(comment).getElement());
-      });
-
-      document.addEventListener(`keydown`, onEscKeyDown);
-
-      popup.getElement()
-        .querySelector(`.film-details__close-btn`)
-        .addEventListener(`click`, (evt) => {
-          evt.preventDefault();
-          unrender(popup.getElement());
-          popup.removeElement();
-        });
-
-      popup.getElement()
-        .querySelector(`.film-details__comment-input`)
-        .addEventListener(`focus`, () => {
-          document.removeEventListener(`keydown`, onEscKeyDown);
-        });
-
-      popup.getElement()
-        .querySelector(`.film-details__comment-input`)
-        .addEventListener(`blur`, () => {
-          document.addEventListener(`keydown`, onEscKeyDown);
-        });
-    };
-
-    card.getElement().querySelector(`.film-card__poster`).addEventListener(`click`, onClickCard);
-    card.getElement().querySelector(`.film-card__title`).addEventListener(`click`, onClickCard);
-    card.getElement().querySelector(`.film-card__comments`).addEventListener(`click`, onClickCard);
-
-    render(container, card.getElement());
+    const movieController = new MovieController(container, filmCard);
+    movieController.init();
   }
 
   _onSortLinkClick(evt) {
